Added widget to display sum of profit/loss for trades
Issue: https://forum.portfolio-performance.info/t/anzahl-trades-um-gewinn-verlust-summe-erweitern/5412
