+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/Messages.java
@@ -607,6 +607,7 @@ public class Messages extends NLS
     public static String LabelTotalValuePercent;
     public static String LabelTrades;
     public static String LabelTradesBasicStatistics;
+    public static String LabelTradesProfitLoss;
     public static String LabelTransactions;
     public static String LabelTransferals;
     public static String LabelTransfer;
