+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages.properties
@@ -1221,6 +1221,8 @@ LabelTrades = Trades
 
 LabelTradesBasicStatistics = Number of trades (with profit/with loss)
 
+LabelTradesProfitLoss = Trades Profit/Loss
+
 LabelTransactions = Transactions
 
 LabelTransfer = Transfer
