+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/messages_de.properties
@@ -1215,6 +1215,8 @@ LabelTrades = Trades
 
 LabelTradesBasicStatistics = Anzahl Trades (mit Gewinn/mit Verlust)
 
+LabelTradesProfitLoss = Trades Gewinn/Verlust
+
 LabelTransactions = Buchungen
 
 LabelTransfer = Umbuchung
